Breaking everything. Reworking everything to load in via json. This requires loading in the products initially and implmenting a three stage build. PreCreateObject, CreateObject and AfterCreateObject. This way everything will get loaded and then three methods will then be called.
PART 2

+++ Platform Game/src/com/platform/main/GameResources/Object/Players/MovableSprite.java
@@ -299,7 +299,7 @@ public abstract class MovableSprite extends AnimatedGameObject
     /*********************************************************************************************/
     @Override
     public void preCreateObject() {
-        this.mTiledTextureRegion = this.gameManager.getMaterialManager().getTiledTexture(this.image, this.width*this.columns, this.height*this.rows, this.columns, this.rows);
+        this.mTiledTextureRegion = this.gameManager.getMaterialManager().getTiledTexture(this.image, ((int)this.width)*this.columns,  ((int)this.height)*this.rows, this.columns, this.rows);
         this.bodyType = BodyDef.BodyType.StaticBody;
     }
